requirements.txt
# Core packages
python-dotenv>=1.0.1
rich>=13.7.0
pyyaml>=6.0.1
requests>=2.32.5
httpx>=0.27.0
urllib3>=2.0.0,<3.0.0
charset-normalizer>=3.3.2
aiohttp>=3.9.5
aiohttp-socks>=0.8.4

# Market data
pandas>=2.1.0
yfinance>=0.2.66

# Table formatting
tabulate>=0.9.0

# Web scraping and text processing
beautifulsoup4>=4.12.0
brotli>=1.1.0
# requests emits a runtime dependency warning with chardet>=6
chardet<6
lxml>=4.9.0
readability-lxml>=0.8.1
tiktoken>=0.7.0

# Web search
# ddgs versions have rapidly changed; ensure we use a release
# compatible with primp. the current known-good version is 9.x.
ddgs>=9.0,<10.0
# primp is used as underlying HTTP client for ddgs and yfinance
primp>=1.1.1
cachetools>=5.0.0
diskcache>=5.6.3

# AI Providers and API server
fastapi>=0.100.0
pydantic>=2.0.0
uvicorn>=0.22.0

# RAG / Vector Memory
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Development / linting
black>=24.8.0
flake8>=7.1.1
pytest>=8.3.0
pytest-asyncio>=0.24.0


